+++ packages/video-processing/src/Exceptions/VideoProcessingFailedException.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace OneOffTech\VideoProcessing\Exceptions;
+
+use Exception;
+
+class VideoProcessingFailedException extends Exception
+{
+    public function __construct($message = '')
+    {
+        parent::__construct($message);
+    }
+}